Stock Transaction getBalance implemented
Provided an endpoint for get balance of stockable product.  Throws a StockableProductNotFoundException if an invalid is passed to the endpoint.

+++ src/main/java/com/eep/stocker/controllers/rest/StockTransactionController.java
@@ -59,11 +59,11 @@ public class StockTransactionController {
     }
 
     @GetMapping("/api/stock-transaction/balance/{id}")
-    public Integer getBalanceForStockableProduct(@PathVariable long id) {
+    public Double getBalanceForStockableProduct(@PathVariable long id) {
         log.info("get: /api/stock-transaction/balance/" + id + " called");
         Optional<StockableProduct> product = stockableProductService.getStockableProductByID(id);
         if(product.isPresent()) {
-            int balance = stockTransactionService.getStockTransactionBalanceForStockableProduct(product.get());
+            double balance = stockTransactionService.getStockTransactionBalanceForStockableProduct(product.get());
             return balance;
         } else {
             throw new StockableProductDoesNotExistException("Stockable product with id of " + id + " does not exist");
